This is the result of a long investigation to accelerate the building of MFE apps in Open edX. Many of the intermediate results can be found in this issue: https://github.com/openedx/wg-frontend/issues/184

The tl;dr is that building MFEs takes a very long time, and that's because of Webpack. In this PR, we replace Webpack by [rsbuild](https://rsbuild.dev/), resulting in a considerable performance improvement.

This PR is dependent on the following upstream PRs:

- https://github.com/openedx/frontend-app-communications/pull/223
- https://github.com/openedx/frontend-app-gradebook/pull/446
- https://github.com/openedx/frontend-app-ora-grading/pull/408
- https://github.com/openedx/frontend-app-learner-dashboard/pull/598

## Open questions

- **Does this still support env.config.jsx customisations?** I think it should, but I haven't checked.
- **Does this build Paragon?** I have no idea. Probably not. To this day I still don't know what's Paragon.

## TODO

Things we need to address before (if?) we merge:

- Remove references to regisb/rsbuild in plugin.py
- Pin rsbuild dependencies in Dockerfile

## Benchmarks

### Configuration

These benchmarks are performed on my personal computer: a Intel(R) Core(TM) i5-4670K CPU @ 3.40GHz with 4 processors and 20 GB RAM. My custom Docker builder runs with `max-parallelism = 4`, as per [these instructions](https://docs.tutor.edly.io/troubleshooting.html#high-resource-consumption-by-docker-on-tutor-images-build):

```
$ docker buildx ls
NAME/NODE        DRIVER/ENDPOINT                           STATUS    BUILDKIT   PLATFORMS
max4cpu*         docker-container                                               
 \_ max4cpu0      \_ unix:///var/run/docker.sock           running   v0.13.1    linux/amd64 (+3), linux/386
```

All plugins are disabled, except for "mfe" and "forum", and there is no bind-mount:

```
tutor plugins disable all
tutor plugins enable mfe forum
tutor mounts list
```

The "mfe" image is built from scratch on the "main" branch; then, we re-build the image by clearing the cache for the "*-prod" build steps, which are the ones that run `npm run build`:

```
# Update environment
tutor config save
# Fill cache
tutor images build mfe
# Build without cache
time tutor images build --docker-arg=--no-cache-filter=account-prod,authn-prod,authoring-prod,communications-prod,discussions-prod,ecommerce-prod,gradebook-prod,learner-dashboard-prod,learning-prod,ora-grading-prod,payment-prod,profile-prod mfe
```

The changes in this PR are tested by checking out this branch:

```
cd <path to>/tutor-mfe
git checkout regisb/rsbuild
```

Then run the commands above.

We list the benchmarks by date, because we expect these figures to change as we iterate on this PR.

### Reference times

#### Reference 1: total build time

On the main branch, the total elapsed ("real") time of that last step is **468 s** (7 min 48 s). This will serve as a point of reference for all benchmarks.

#### Reference 2: no parallelism build of the learning MFE

Another interesting point of reference is the time that is needed to build the learning MFE, without parallelism -- that is, without concurrent builds of other MFEs. This is a figure that can be obtained by running:

```
time tutor images build --docker-arg=--no-cache-filter=learning-prod --target=learning-prod mfe
```

And then checking the time to build the MFE in the Docker logs: 

```
 => [learning-prod 1/1] RUN npm run build                                119.5s
```

Here, the reference time is **119.5s**.

### Single-thread build of the learning app

### 2025/04/09 Initial results

- **Ref 1:** **105s** (4.46x improvement)
- **Ref 2:** **7.1s** (14.8x improvement)

All MFEs build under 23s, despite the fact that there are 4 builds in parallel:

```
 => [communications-prod 2/2] RUN rsbuild build                                       21.7s
 => [learner-dashboard-prod 2/2] RUN rsbuild build                                    23.0s
 => [learning-prod 2/2] RUN rsbuild build                                             22.7s
 => [discussions-prod 2/2] RUN rsbuild build                                          22.6s
 => [gradebook-prod 2/2] RUN rsbuild build                                            24.1s
 => [ora-grading-prod 2/2] RUN rsbuild build                                          22.6s
 => [account-prod 2/2] RUN rsbuild build                                              22.7s
 => [authn-prod 2/2] RUN rsbuild build                                                22.7s
 => [profile-prod 2/2] RUN rsbuild build                                              18.6s
 => [authoring-prod 2/2] RUN rsbuild build                                            18.5s
```