@regisb 2025-04-10
> Is there anything you found that doesn't work / isn't supported by rsbuild?

No, I haven't. That's because there's nothing magical with our MFEs: they are just frontend code, bundled as regular js/html/css/images etc. 

That being said, I haven't extensively tested the 10 officially maintained MFEs, so it's quite possible I missed some issues.

> If not, we should just replace it everywhere and not just for tutor-mfe.

I'm not sure anymore about that. While I worked on this project, I discovered [some](https://github.com/openedx/frontend-app-discussions/blob/60da5eafc4138d131097013ee957deb37dfcc144/src/components/TinyMCEEditor.jsx#L38) [places](https://github.com/openedx/frontend-app-learning/blob/85e6e9266d15ad5d5922e20ae767383da7f05489/jest.config.js#L15) [where](https://github.com/openedx/frontend-app-learning/blob/85e6e9266d15ad5d5922e20ae767383da7f05489/.eslintrc.js#L18) the webpack abstraction leaked into the MFE codebase, and that really feels like an anti-pattern.

Instead, I now think that we should remove all of the build code from all MFE. There shouldn't be any MFE-specific build step. There should be a simple, documented standard to build any MFE. And we should be able to run something like `mfe-build --dev ./frontend-app-learning`. This "mfe-build" tool could use rsbuild under the hood, or something else, and we should be able to replace that bundler easily in the future, when a better, shinier bundler inevitably comes out.