@hinakhadim 2025-04-10
Awesome! I tested this PR on my system, and it significantly reduces the build time—from 2007 seconds (using npm run build / Webpack) to 754 seconds (using rsbuild).

**With Webpack**: `[+] Building 2007.7s (127/127) FINISHED                     docker-container:doublecpu`
<img width="1784" alt="image" src="https://github.com/user-attachments/assets/4a96cac6-04fa-404b-b7de-51219e618b92" />


**With Rsbuild:**  `[+] Building 754.1s (138/138) FINISHED                        docker-container:doublecpu`
<img width="1784" alt="image" src="https://github.com/user-attachments/assets/743c5077-73ac-497c-b037-cd2d59a8011a" />

I am trying to start(tutor local launch) the system and will give more feedback after `tutor local launch`.