@bradenmacdonald 2025-04-10
> Instead, I now think that we should remove all of the build code from all MFE. There shouldn't be any MFE-specific build step. There should be a simple, documented standard to build any MFE. And we should be able to run something like mfe-build --dev ./frontend-app-learning. This "mfe-build" tool could use rsbuild under the hood, or something else, and we should be able to replace that bundler easily in the future, when a better, shinier bundler inevitably comes out.

This makes sense to me, and I think it may even have been the original idea behind `frontend-build`, but things have drifted away from that a bit over time.