@regisb 2025-04-14
Thanks for testing @Danyal-Faheem! Would you mind trying out debugging with `RUST_BACKTRACE`, as suggested in the error logs?