@DawoudSheraz 2025-04-16
I tested this branch on Macbook M2 and got the same error as Danyal mentioned above https://github.com/overhangio/tutor-mfe/pull/252#issuecomment-2796362845. However, even when I specified the platform (amd/64) in build argument, the build process did not complete and gave the same error. npm connection errors were happening as well, which has been more prominent than other build issues for me (even though I am using max2cpu config). I am adding some stats below:

- Indigo enabled
  - Mfe build with —no-cache
    - 614.9s, failed due to storage size issue
  - Mfe build —no-cache —no-registry-cache
    - Failed due to react compatibility issues in Indigo, 1115.7 seconds
- Indigo Disabled
  - Mfe build —no-cache —no-registry-cache
      - Failed with network error, 795.1s
  - Re-tried but no cache
    - Failed with rsbuild error, 403.1 seconds