@regisb 2025-04-22
I was able to reproduce the issue by building on a remote arm64 server:

```
$ tutor images build --docker-arg=--platform=arm64 --docker-arg=--target=learner-dashboard-prod mfe
...
0.427 
0.427   Rsbuild v1.3.9
0.427 
1.165 info    Rspack persistent cache enabled (experimental)
9.090 Panic occurred at runtime. Please file an issue on GitHub with the backtrace below: https://github.com/web-infra-dev/rspack/issues
9.090 Message:  created a new `Panic` from: out of range integral type conversion attempted
9.090 Location: index.crates.io-1949cf8c6b5b557f/rancor-0.1.0/src/lib.rs:648
9.090 
9.090 Backtrace omitted.
9.090 
9.090 Run with RUST_BACKTRACE=1 environment variable to display it.
9.090 Run with RUST_BACKTRACE=full to include source snippets.
9.254 Aborted (core dumped)
```

The issue disappears when I disable the cache, which is tagged as "experimental". So what I did was to disable caching on arm64. I also filed the following upstream issue: https://github.com/web-infra-dev/rspack/issues/10118

Thus, Apple users should now be able to test this PR again. :eyes: @Danyal-Faheem @DawoudSheraz