@Danyal-Faheem 2025-04-23
@regisb I tested it with the updated commits and It's building on arm64 natively now!

The results are also impressive:

Running with a max-parallelism of 2 (I had to do this otherwise the `npm clean-install` step fails due to network issues), The rsbuild steps are 3-4x faster on average than npm build.

Some stats:

### Build Time Comparison

| Component           | Webpack/NPM (s) | rsbuild (s) |
|--------------------|--------------------|-----------------------------|
| Full MFE Build     | 1228.1             | 748.7                       |
| Learner-Dashboard  | 111.9              | 48.5                        |
| Learning           | 136.3              | 11.1                        |
| Profile            | 86.0               | 13.9                        |